Implemented Auger top-limit switch and overshooting fix.
The Auger now stops when the upper limit switch is tripped by the top-most
frisbee. Also, the overshoot bug which required the rotate command to be
called twice when rotating in a different direction than the previous one,
is fixed. This is done by storing the previous direction and checking if
the current direction is different to that, and if so, rotating for two
pulses.